+++ src/plugins/widgets/search/types.ts
@@ -11,7 +11,7 @@ type Data = {
 export type Props = API<Data>;
 
 export const defaultData: Data = {
-  searchEngine: "google",
+  searchEngine: "default",
   suggestionsQuantity: 4,
 };